Release 2.2.0

+++ app/modules/AmazingCat/ProjectsModule/module.init.js
@@ -1,14 +1,14 @@
 import moment from 'moment';
 import 'moment-timezone';
-import ProjectService from '../../../core/service/resource/projectService';
-import i18n from '../../../core/i18n';
-import UserAvatar from '../../../core/components/UserAvatar';
-import { havePermission } from '@/utils/user';
+import ProjectService from '@/service/resource/projectService';
+import i18n from '@/i18n';
+import UserAvatar from '@/components/UserAvatar';
+import { formatDurationString } from '@/utils/time';
+import { ModuleLoaderInterceptor } from '@/moduleLoader';
 
 export const ModuleConfig = {
     enabled: true,
     routerPrefix: 'projects',
-    initOrder: 4
 };
 
 function formatDateTime(value, timezone) {
@@ -17,6 +17,23 @@ function formatDateTime(value, timezone) {
 }
 
 export function init(context, router) {
+    let routes = {};
+    ModuleLoaderInterceptor.on('AmazingCat_CoreModule', m => {
+        m.routes.forEach(route => {
+            if (route.name.search('users.view') > 0) {
+                routes.usersView = route.name;
+            }
+        });
+    });
+
+    ModuleLoaderInterceptor.on('AmazingCat_TasksModule', m => {
+        m.routes.forEach(route => {
+            if (route.name.search('view') > 0) {
+                routes.tasksView = route.name;
+            }
+        });
+    });
+
     const crud = context.createCrud('projects.crud-title', 'projects', ProjectService);
     crud.view.addToMetaProperties('titleCallback', ({ values }) => values.name, crud.view.getRouterConfig());
 
@@ -29,67 +46,133 @@ export function init(context, router) {
 
     const grid = context.createGrid('projects.grid-title', 'projects', ProjectService, {
         with: ['users'],
-        withCount: ['tasks']
+        withCount: ['tasks'],
     });
 
     const fieldsToShow = [
         {
             label: 'field.name',
-            key: 'name'
+            key: 'name',
         },
         {
             label: 'field.created_at',
             key: 'created_at',
             render: (h, { currentValue, companyData }) => {
                 return h('span', formatDateTime(currentValue, companyData.timezone));
-            }
+            },
         },
         {
             label: 'field.updated_at',
             key: 'updated_at',
             render: (h, { currentValue, companyData }) => {
                 return h('span', formatDateTime(currentValue, companyData.timezone));
-            }
+            },
         },
         {
             label: 'field.description',
-            key: 'description'
-        }
+            key: 'description',
+        },
+        {
+            key: 'total_spent_time',
+            label: 'field.total_spent',
+            render: (h, props) => h('span', formatDurationString(props.currentValue)),
+        },
+        {
+            key: 'workers',
+            label: 'field.workers',
+            render: (h, props) => {
+                const data = [];
+                Object.keys(props.currentValue).forEach(k => {
+                    props.currentValue[k].time = formatDurationString(+props.currentValue[k].duration);
+                    data.push(props.currentValue[k]);
+                });
+                return h('AtTable', {
+                    props: {
+                        columns: [
+                            {
+                                title: 'User',
+                                render: (h, { item }) => {
+                                    return h(
+                                        'router-link',
+                                        {
+                                            props: {
+                                                to: {
+                                                    name: routes.usersView,
+                                                    params: { id: item.user_id },
+                                                },
+                                            },
+                                        },
+                                        item.full_name,
+                                    );
+                                },
+                            },
+                            {
+                                title: 'Task Name',
+                                render: (h, { item }) => {
+                                    return h(
+                                        'router-link',
+                                        {
+                                            props: {
+                                                to: {
+                                                    name: routes.tasksView,
+                                                    params: { id: item.task_id },
+                                                },
+                                            },
+                                        },
+                                        item.task_name,
+                                    );
+                                },
+                            },
+                            {
+                                key: 'time',
+                                title: 'Time',
+                                render(h, { item }) {
+                                    return h('div', {
+                                        domProps: {
+                                            textContent: !item ? '0h 0m' : item.time,
+                                        },
+                                        styles: {
+                                            'white-space': 'nowrap',
+                                        },
+                                    });
+                                },
+                            },
+                        ],
+                        data,
+                        pagination: true,
+                        'page-size': 100,
+                    },
+                });
+            },
+        },
     ];
 
     const fieldsToFill = [
         {
             key: 'id',
-            displayable: false
+            displayable: false,
         },
         {
             label: 'field.name',
             key: 'name',
             type: 'text',
             placeholder: 'field.name',
-            required: true
+            required: true,
         },
         {
             label: 'field.description',
             key: 'description',
             type: 'textarea',
-            required: true
+            required: true,
+            placeholder: 'field.description',
         },
         {
             label: 'field.important',
+            tooltipValue: 'tooltip.task_important',
             key: 'important',
-            type: 'select',
-            options: [
-                {
-                    value: 0,
-                    label: 'control.no'
-                },
-                {
-                    value: 1,
-                    label: 'control.yes'
-                }
-            ]
-        }
+            type: 'checkbox',
+            default: 0,
+        },
     ];
 
     crud.view.addField(fieldsToShow);
@@ -99,7 +182,7 @@ export function init(context, router) {
     grid.addColumn([
         {
             title: 'field.project',
-            key: 'name'
+            key: 'name',
         },
         {
             title: 'field.team',
@@ -107,22 +190,30 @@ export function init(context, router) {
             render: (h, { item }) => {
                 const users = item.users || [];
 
-                return h('div', { class: 'projects-grid__initials-row' }, users.map(user => {
-                    return h('AtTooltip', {
-                        props: {
-                            placement: 'top',
-                            content: user.full_name
-                        }
-                    }, [
-                        h(UserAvatar, {
-                            props: {
-                                user,
-                                showTooltip: true
-                            }
-                        })
-                    ])
-                }));
-            }
+                return h(
+                    'div',
+                    { class: 'projects-grid__initials-row' },
+                    users.map(user => {
+                        return h(
+                            'AtTooltip',
+                            {
+                                props: {
+                                    placement: 'top',
+                                    content: user.full_name,
+                                },
+                            },
+                            [
+                                h(UserAvatar, {
+                                    props: {
+                                        user,
+                                        showTooltip: true,
+                                    },
+                                }),
+                            ],
+                        );
+                    }),
+                );
+            },
         },
         {
             title: 'field.amount_of_tasks',
@@ -130,24 +221,22 @@ export function init(context, router) {
             render: (h, { item }) => {
                 const amountOfTasks = item.tasks_count || 0;
 
-                return h('span', i18n.tc(
-                    'projects.amount_of_tasks',
-                    amountOfTasks,
-                    {
-                        count: amountOfTasks
-                    }
-                ));
-            }
-        }
+                return h(
+                    'span',
+                    i18n.tc('projects.amount_of_tasks', amountOfTasks, {
+                        count: amountOfTasks,
+                    }),
+                );
+            },
+        },
     ]);
 
     grid.addFilter([
         {
             referenceKey: 'name',
-            filterName: 'filter.project'
-        }
+            filterName: 'filter.fields.project_name',
+        },
     ]);
-
 
     // TODO when assign users will be ready -- uncomment it
     // const assignRouteName = context.getModuleRouteName() + '.assign';
@@ -171,7 +260,7 @@ export function init(context, router) {
             },
             renderCondition: () => {
                 return false; // TODO
-            }
+            },
         },
         {
             title: 'control.view',
@@ -182,7 +271,7 @@ export function init(context, router) {
             renderCondition({ $store }) {
                 // User always can view assigned projects
                 return true;
-            }
+            },
         },
         {
             title: 'control.edit',
@@ -190,42 +279,69 @@ export function init(context, router) {
             onClick: (router, params) => {
                 router.push({ name: crudEditRoute, params: { id: params.item.id } });
             },
-            renderCondition: ({ $store }) => {
-                return havePermission($store.getters['user/allowedRules'], 'projects/edit');
-            }
+            renderCondition: ({ $store }, item) => {
+                return $store.getters['user/can']('projects/edit', item.id);
+            },
         },
         {
             title: 'control.delete',
             actionType: 'error', // AT-UI action type,
             icon: 'icon-trash-2',
             onClick: async (router, params, context) => {
-                const projectService = new ProjectService();
-                await projectService.deleteItem(params.item.id);
-                context.tableData = context.tableData.filter(item => item.id !== params.item.id);
-                context.$Notify({
-                    type: 'success',
-                    title: 'Success',
-                    message: 'Project deleted successfully'
+                const res = await context.$CustomModal({
+                    title: 'Delete this Project?',
+                    content: 'After deletion this project cannot be restored',
+                    okText: 'Delete',
+                    cancelText: 'Cancel',
+                    showClose: false,
+                    styles: {
+                        'border-radius': '10px',
+                        'text-align': 'center',
+                        footer: {
+                            'text-align': 'center',
+                        },
+                        header: {
+                            padding: '16px 35px 4px 35px',
+                            color: 'red',
+                        },
+                        body: {
+                            padding: '16px 35px 4px 35px',
+                        },
+                    },
+                    width: 320,
+                    type: 'trash',
+                    typeButton: 'error',
                 });
+
+                if (res === 'confirm') {
+                    const projectService = new ProjectService();
+                    await projectService.deleteItem(params.item.id);
+                    context.tableData = context.tableData.filter(item => item.id !== params.item.id);
+                    context.$Notify({
+                        type: 'success',
+                        title: 'Success',
+                        message: 'Project deleted successfully',
+                    });
+                }
             },
-            renderCondition: ({ $store }) => {
-                return havePermission($store.getters['user/allowedRules'], 'projects/remove');
-            }
-        }
+            renderCondition: ({ $store }, item) => {
+                return $store.getters['user/can']('projects/remove', item.id);
+            },
+        },
     ]);
 
     grid.addPageControls([
         {
             label: 'control.create',
             renderCondition: ({ $store }) => {
-                return havePermission($store.getters['user/allowedRules'], 'projects/create');
+                return $store.getters['user/can']('projects/create');
             },
             type: 'primary',
             icon: 'icon-edit',
             onClick: ({ $router }) => {
                 $router.push({ name: crudNewRoute });
-            }
-        }
+            },
+        },
     ]);
 
     context.addRoute(crud.getRouterConfig());
@@ -234,13 +350,13 @@ export function init(context, router) {
     context.addNavbarEntry({
         label: 'navigation.projects',
         to: {
-            name: 'AmazingCat_ProjectsModule.crud.projects'
-        }
+            name: 'AmazingCat_ProjectsModule.crud.projects',
+        },
     });
 
     context.addLocalizationData({
         en: require('./locales/en'),
-        ru: require('./locales/ru')
+        ru: require('./locales/ru'),
     });
 
     return context;
